refactor: utils, actions 분리
